Move plan label normalization to a helper module
At some point, Linode plan labels migrated from including the plan
size in bytes without unit (e.g., `2048`) to including the plan
size in gigabytes with the unit (e.g., `2GB`).

The previous attempt to transparently handle byte labels is updated
to move the label-munging logic to a reusable helper module so that
the functionality is implemented consistently wherever it is used.